% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addTooltip.R
\name{addTooltip}
\alias{addTooltip}
\title{addTooltip}
\usage{
addTooltip(session, id, title, placement = "bottom", trigger = "hover",
  options = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{id}{The id of the element to attach the tooltip to.}

\item{title}{The content of the tooltip.}

\item{placement}{Where the tooltip should appear relative to its target
(\code{top}, \code{bottom}, \code{left}, or \code{right}). Defaults to \code{"bottom"}.}

\item{trigger}{What action should cause the tooltip to appear? (\code{hover},
\code{focus}, \code{click}, or \code{manual}). Defaults to \code{"hover"}.}

\item{options}{A named list of additional options to be set on the tooltip.}
}
\description{
\code{addTooltip} is used within the Server logic of an app to add a tooltip to a Shiny
input or output.
}
\details{
See \link{Tooltips_and_Popovers} for more information about how to use \code{addTooltip} with the
rest of the Tooltips_and_Popovers family.
}
\note{
Run \code{bsExample("Tooltips_and_Popovers")} for an example
of \code{addTooltip} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Tooltips_and_Popovers: \code{\link{Tooltips_and_Popovers}};
  \code{\link{addPopover}}; \code{\link{bsPopover}};
  \code{\link{bsTooltip}}; \code{\link{popify}};
  \code{\link{removePopover}}; \code{\link{removeTooltip}};
  \code{\link{tipify}}
}

