% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsAlert.R
\name{Alerts}
\alias{Alerts}
\alias{bsAlert}
\alias{closeAlert}
\alias{createAlert}
\title{Alerts}
\usage{
bsAlert(anchorId)

createAlert(session, anchorId, alertId = NULL, title = NULL,
  content = NULL, style = NULL, dismiss = TRUE, append = TRUE)

closeAlert(session, alertId)
}
\arguments{
\item{anchorId}{A unique ID that identifies the anchor.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{alertId}{A unique ID that identifies the specific alert being created.}

\item{title}{An optional title for the alert. This will appear at the top of the alert in larger font.}

\item{content}{The main body of the alert. HTML tags are allowed.}

\item{style}{The Bootstrap style to apply (\code{danger}, \code{warning}, \code{info}, or \code{success}.}

\item{dismiss}{\code{logical} indicating whether the alert should be user dismissable.}

\item{append}{\code{logical} indicating whether the alert should be appended to the anchor, below any existing alerts.}
}
\description{
Alerts allow you to communicate information to the user on the fly. Standard
Bootstrap styling options give the user a hint at the type of information
contained in the Alert.
}
\details{
To create alerts in your Shiny app you must place \code{bsAlert} in your ui.
This serves as an anchor that tells shinyBS where to place the alerts created
with \code{createAlert}.

Use \code{createAlert} in your server script to add alerts to the anchor
you created with \code{bsAlert} in your ui. You can place \code{createAlert}
in observers, reactives, or outputs. A common usage may be to have logic that
validates a user's inputs. If they are valid produce the requested output, if
not use \code{createAlert} to give the user info about what they need to
change.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\section{Changes}{

\code{style} was called \code{type} in previous versions of shinyBS.

\code{anchorId} was called \code{inputId} in previous versions of shinyBS.

\code{content} was called \code{message} in previous versions of shinyBS.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}
}

