% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsAlert.R
\name{bsAlert}
\alias{bsAlert}
\title{bsAlert}
\usage{
bsAlert(anchorId)
}
\arguments{
\item{anchorId}{A unique id the identifies the anchor.}
}
\description{
\code{bsAlert} creates an anchor point in your UI definition. This anchor point
is where alerts created in your Server logic will be displayed.
}
\details{
See \link{Alerts} for more information about how to use \code{bsAlert} with the
rest of the Alerts family.
}
\note{
Run \code{bsExample("Alerts")} for an example
of \code{bsAlert} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Alerts: \code{\link{Alerts}};
  \code{\link{closeAlert}}; \code{\link{createAlert}}
}

