% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsButton.R
\name{bsButton}
\alias{bsButton}
\title{bsButton}
\usage{
bsButton(inputId, label, icon = NULL, ..., style = "default",
  size = "default", type = "action", block = FALSE, disabled = FALSE,
  value = FALSE)
}
\arguments{
\item{inputId}{Specifies the input slot that will be used to access the
value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link{icon}} to appear on the button.}

\item{...}{Named attributes to be applied to the button or link.}

\item{style}{A Bootstrap style to apply to the button. (\code{default}, \code{primary},
\code{success}, \code{info}, \code{warning}, or \code{danger})}

\item{size}{The size of the button (\code{extra-small}, \code{small},
\code{default}, or \code{large})}

\item{type}{The type of button to create. (\code{action} or \code{toggle})}

\item{block}{\bold{logical} Should the button take the full width of the parent element?}

\item{disabled}{\bold{logical} Should the button be disabled (un-clickable)?}

\item{value}{\bold{logical} If \code{type = "toggle"}, the initial value of the button.}
}
\description{
\code{bsButton} is used in your UI script to create customizable action and toggle
buttons.
}
\details{
See \link{Buttons} for more information about how to use \code{bsButton} with the
rest of the Buttons family.
}
\note{
Run \code{bsExample("Buttons")} for an example
of \code{bsButton} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Buttons: \code{\link{Buttons}};
  \code{\link{updateButton}}
}

