% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsCollapse.R
\name{bsCollapse}
\alias{bsCollapse}
\title{bsCollapse}
\usage{
bsCollapse(..., id = NULL, multiple = FALSE, open = NULL)
}
\arguments{
\item{id}{If provided, you can use input$id in your server script to open or
close \code{bsCollapsePanels} within the \code{bsCollapse}.}

\item{multiple}{\code{logical} indication whether multiple \code{bsCollapsePanels}
within this \code{bsCollapse} can be open at once.}

\item{open}{The \code{value} (or, if none was supplied, the title) of the
\code{bsCollapsePanel} that should be open. If \code{multiple = TRUE}, then
this can be a vector of \code{value}s or \code{title}s.}

\item{\dots}{\code{\link{bsCollapsePanel}} elements to include in the \code{bsCollapse}.}
}
\description{
Create a collapse that contains \code{\link{bsCollapsePanel}} elements.
See \code{\link{Collapses}}
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}
}
\seealso{
Other Collapses: \code{\link{Collapses}};
  \code{\link{bsCollapsePanel}};
  \code{\link{updateCollapse}}
}

