% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsCollapseButton.R
\name{bsCollapseButton}
\alias{bsCollapseButton}
\alias{bsCollapseLink}
\title{Create a button or link that toggles the state of a
\code{\link{bsCollapsePanel}}.}
\usage{
bsCollapseButton(
  inputId,
  label,
  target,
  container = NULL,
  icon = NULL,
  width = NULL,
  ...
)

bsCollapseLink(
  inputId,
  label,
  target,
  parent = NULL,
  icon = NULL,
  width = NULL,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{target}{HTML id of object to be toggled}

\item{icon}{An optional \code{\link[shiny:icon]{icon()}} to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:reexports]{validateCssUnit()}}.}

\item{...}{Named attributes to be applied to the button or link.}

\item{parent}{\bold{Optional} HTML id of parent container}
}
\description{
Create a button or link that toggles the state of a
\code{\link{bsCollapsePanel}}.
}
\details{
See \link{Collapses} for more information about how to use \code{bsCollapseButton} with the
rest of the Collapses family.
}
\section{Functions}{
\itemize{
\item \code{bsCollapseButton}: Create an \code{\link[shiny]{actionLink}}
that will toggle the collapsed state of a
\code{\link{bsCollapsePanel}} when clicked.

\item \code{bsCollapseLink}: Create an \code{\link[shiny]{actionLink}}
that will toggle the collapsed state of a
\code{\link{bsCollapsePanel}} when clicked.
}}

\note{
Run \code{bsExample("Collapses")} for an example
of \code{bsCollapseButton} functionality.

Providing `parent` causes all other collapsible elements under the
 specified parent to be closed, so that no more than one panel is open
 at a time.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Collapses: 
\code{\link{Collapses}},
\code{\link{bsCollapsePanel}()},
\code{\link{bsCollapse}()},
\code{\link{updateCollapse}()}
}
\concept{Collapses}
