% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsCollapse.R
\name{bsCollapsePanel}
\alias{bsCollapsePanel}
\title{bsCollapsePanel}
\usage{
bsCollapsePanel(title, ..., value = title, style = NULL)
}
\arguments{
\item{title}{The title to display at the top of the \code{bsCollapsePanel}}

\item{value}{The value that should be returned to the server when this
\code{bsCollapsePanel} is open.}

\item{style}{A Bootstrap style to apply (\code{danger}, \code{warning}, \code{info}, or \code{success}.}

\item{\dots}{UI elements to include within the collapse panel}
}
\description{
Create a collapse panel that can be included within a \code{\link{bsCollapse}}
See \code{\link{Collapses}}
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}
}
\seealso{
Other Collapses: \code{\link{Collapses}};
  \code{\link{bsCollapse}}; \code{\link{updateCollapse}}
}

