% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsExample.R
\name{bsExample}
\alias{bsExample}
\title{bsExample}
\usage{
bsExample(family, display.mode = "showcase", ...)
}
\arguments{
\item{family}{A shinyBS family name}

\item{display.mode}{The display mode to use when running the example. See
\code{\link{runApp}}.}

\item{\dots}{Other parameters to pass to \code{\link{runApp}}.}
}
\description{
A function to view examples of shinyBS functionality. Will run the examples
found in the examples sections of shinyBS documentation. Use this instead of
\code{example}.
}
\details{
This function is just a wrapper for \code{\link{runApp}} that runs copies of the
examples found in the family documention pages of \code{shinyBS}. By default,
\code{display.mode} is set to \code{showcase} so you can see the code while
the app is running.
}
\examples{
\dontrun{
   bsExample("Alerts")}
}

