% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bsMenu}
\alias{bsMenu}
\title{Menu Dropdowns}
\usage{
bsMenu(menuId, ..., label, icon = NULL, type = "toolbar", caret = TRUE,
  target = NULL)
}
\arguments{
\item{menuId}{An Id for the menu}

\item{\dots}{Items to add to the menu}

\item{label}{The label for the menu}

\item{icon}{An optional Font Awesome icon}

\item{type}{The type of toolbar to create; either toolbar, submenu, or popup
(See details)}

\item{caret}{Logical indicating whether to include a caret on a toolbar menu}

\item{target}{Id of a target for the popup menu}
}
\description{
Menu Dropdowns
}
\details{
Function for creating various menu types
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/components.html}{Components of
 Twitter Bootstrap 2.3.2}

\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}
}

