% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bsMenuDateRangeInput}
\alias{bsMenuDateRangeInput}
\title{Date Range Input for shinyBS menu bars}
\usage{
bsMenuDateRangeInput(inputId, label, start = NULL, end = NULL, min = NULL,
  max = NULL, format = "yyyy-mm-dd", startview = "month", weekstart = 0,
  language = "en", presets)
}
\arguments{
\item{inputId}{An Id for the dateRangeInput object}

\item{label}{Label for dateRangeInput (appears as tooltip)}

\item{start}{See \link{dateRangeInput}}

\item{end}{See \link{dateRangeInput}}

\item{min}{See \link{dateRangeInput}}

\item{max}{See \link{dateRangeInput}}

\item{format}{See \link{dateRangeInput}}

\item{startview}{See \link{dateRangeInput}}

\item{weekstart}{See \link{dateRangeInput}}

\item{language}{See \link{dateRangeInput}}

\item{presets}{A list of preset date ranges to aid in date selection (See
 details)}
}
\description{
Function for creating a date range input for shinyBS menubars
}
\details{
\code{bsMenuDateRangeInput} is a modified version of Shiny's native
 \code{\link{dateRangeInput}} designed to fit better within a shinyBS
 menubar. \cr\cr If \code{presets} is provided, a button with a calendar icon
 will be prepended to the date range input. Clicking this button will reveal
 a dropdown with the provided presets. If the user clicks a preset, the
 dateRangeInput will automatically be updated to selected preset. Presets
 should be in the form of a named list of two-element vectors. For
 example:\cr\cr \code{presets = list("Year Two" = c("2013-01-01",
 "2013-12-31"), "Year One" = c("2012-01-01", "2012-12-31"))} \cr\cr Would
 create two presets labeled 'Year One' and 'Year Two' corresponding to the
 provided date ranges. \cr\cr There is currently no update function for
 \code{bsMenuDateRangeInput}. the standard \code{\link{updateDateRangeInput}}
 will work on \code{bsMenuDateRangeInput} but does not give the option of
 updating the presets. If that is functionality that you would find useful,
 please let contact me.
}
\note{
Run \code{bsDemo("menubar")} for a live examples of shinyBS menubar
 functionality.
}
\examples{
#Run bsDemo("menubar") for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/components.html}{Components of
 Twitter Bootstrap 2.3.2}

\href{http://getbootstrap.com/2.3.2/base-css.html}{Base CSS for
 Twitter Bootstrap 2.3.2}
}

