% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsModal.R
\name{bsModal}
\alias{bsModal}
\title{bsModal}
\usage{
bsModal(id, title, trigger, ..., size, footer = NULL, close.button = TRUE,
  width = NULL)
}
\arguments{
\item{id}{A unique identifier for the modal window}

\item{title}{The title to appear at the top of the modal}

\item{trigger}{The id of a button or link that will open the modal.}

\item{size}{\bold{Optional} What size should the modal be? (\code{small} or \code{large})}

\item{footer}{A \code{list} of shiny UI elements to be added to the footer of the modal.}

\item{close.button}{Should a close button be added to the footer of the modal?}

\item{width}{An optional width argument for the modal. Must include units. Only applied if \code{size} is missing.}

\item{\dots}{UI elements to include within the modal}
}
\description{
\code{bsModal} is used within the UI to create a modal window.
}
\details{
See \link{Modals} for more information about how to use \code{bsModal} with the
rest of the Modals family.
}
\note{
Run \code{bsExample("Modals")} for an example
of \code{bsModal} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Modals: \code{\link{Modals}},
  \code{\link{toggleModal}}
}

