% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsPopover.R
\name{bsPopover}
\alias{bsPopover}
\title{bsPopover}
\usage{
bsPopover(id, title, content, placement = "bottom", trigger = "hover",
  options = NULL)
}
\arguments{
\item{id}{The id of the element to attach the popover to.}

\item{title}{The title of the popover.}

\item{content}{The main content of the popover.}

\item{placement}{Where the popover should appear relative to its target
(\code{top}, \code{bottom}, \code{left}, or \code{right}). Defaults to \code{"bottom"}.}

\item{trigger}{What action should cause the popover to appear? (\code{hover},
\code{focus}, \code{click}, or \code{manual}). Defaults to \code{"hover"}.}

\item{options}{A named list of additional options to be set on the popover.}
}
\description{
\code{bsPopover} is used within the UI of an app to add a popover to a Shiny
input or output.
}
\details{
See \link{Tooltips_and_Popovers} for more information about how to use \code{bsPopover} with the
rest of the Tooltips_and_Popovers family.
}
\note{
Run \code{bsExample("Tooltips_and_Popovers")} for an example
of \code{bsPopover} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Tooltips_and_Popovers: \code{\link{Tooltips_and_Popovers}};
  \code{\link{addPopover}}; \code{\link{addTooltip}};
  \code{\link{bsTooltip}}; \code{\link{popify}};
  \code{\link{removePopover}}; \code{\link{removeTooltip}};
  \code{\link{tipify}}
}

