% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTypeahead.R
\name{bsTypeahead}
\alias{bsTypeahead}
\title{bsTypeahead}
\usage{
bsTypeahead(inputId, label, value = "", choices, items = 8, minLength = 1)
}
\arguments{
\item{inputId}{Input variable to assign the control's value to}

\item{label}{Display label for the control}

\item{value}{Initial value}

\item{choices}{Array of strings to match against. Can also be JavaScript
function. Use htmlwidgets::JS() to indicate JavaScript. The length of \code{choices} must no exceed 131370.}

\item{items}{The max number of items to display in the dropdown. Can also be
set to 'all'}

\item{minLength}{The minimum character length needed before triggering
autocomplete suggestions. You can set it to 0 so suggestion are shown even
when there is no text when lookup function is called.}
}
\description{
\code{typeaheadInput} creates a textinput with type ahead function buttons.
}
\seealso{
\code{\link{updateTypeaheadInput}}
}
