% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsButtons.R
\name{Buttons}
\alias{Buttons}
\alias{bsButton}
\title{Create Twitter Bootstrap Buttons}
\usage{
bsButton(inputId, label, icon = NULL, ..., style = "default", size = "md",
  type = "action", block = FALSE, disabled = FALSE, value = FALSE)
}
\arguments{
\item{inputId}{Input variable to assign the control's value to}

\item{label}{The text that will appear within the button}

\item{style}{The style to apply to the button}

\item{size}{The desired size of the button}

\item{block}{Logical indicating whether the button should be a block level
control.}

\item{disabled}{Logical indicating whether the button should be clickable}

\item{value}{See details}
}
\description{
Functions to create and update Twitter Bootstrap Buttons in shiny
}
\details{


shinyBS adds additional styling options to the
\code{\link{actionButton}} and introduces two new types of buttons to the
shiny interface.

\code{bsButton} is a generic button that doesn't have any default
behavior. They are used as inputs when creating a
\code{\link{bsButtonGroup}}. They can also be used as triggers for other
elements like \code{\link{bsModal}} or \code{\link{bsPopover}}.
}
\note{
Run \code{bsDemo()} for a live example of shinyBS Buttons functionality.
}
\section{Options}{

For \code{bsButton}, \code{value} is used to assign a return value for
\code{\link{bsButtonGroup}}. See \code{\link{bsButtonGroup}} for more
details. For \code{bsToggleButton} and \code{updateButton}, \code{value} is
used to set whether the button is currently 'clicked'.

\code{style} is used to customize the appearence of buttons. There are six
options for style: \code{primary}, \code{info}, \code{success},
\code{warning}, \code{danger}, \code{inverse}, \code{link}. Any other value
will result in a standard button style.

\code{size} is used to control the size of buttons. \code{size} can be
\code{large}, \code{small}, or \code{mini}. Any other value will result in a
standard-sized button.

Setting \code{block = TRUE} will cause the button to take up the full width
of its parent element.

When \code{disabled = TRUE} the button will appear "greyed-out" and clicking
the button will have no effect on the app.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/}{Twitter Bootstrap 2.3.2}
}

