% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Collapse Panels}
\alias{Collapse Panels}
\alias{bsCollapse}
\alias{bsCollapsePanel}
\alias{updateCollapse}
\title{Twitter Bootstrap Collapse Panels}
\usage{
bsCollapse(..., id = NULL, multiple = FALSE, open = NULL)

bsCollapsePanel(title, ..., id = NULL, value = NULL)

updateCollapse(session, id, open = NULL, close = NULL, multiple = NULL)
}
\arguments{
\item{\dots{}}{For \code{bsCollapse} a series of \code{bsCollapsePanels} to
include in the object for \code{bsCollapsePanel} as series of UI elements to
add to the panel}

\item{id}{A name for the object (Optional)}

\item{multiple}{Logical indicating whether multiple panels in a group can be
open at one time}

\item{open}{The id of the panel(s) you want to open}

\item{close}{The id of the panel(s) you want to close}

\item{title}{The title to appear at the top of the panel}

\item{value}{The value to return to bsCollapse when this panel is open}

\item{session}{The \code{session} object passed to function given to
 \code{shinyServer}}
}
\description{
Functions for creating and manipulating Collapse (Accordion)
 Panels in shiny
}
\details{
\code{bsCollapse} works very similarly to \code{tabsetPanel} and may
 be useful in similar situations.\cr\cr Like \code{tabsetPanel}s,
 \code{bsCollapse} can return values indicating which panels are open or
 \code{NULL} if no panels are open.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}
}

