% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsCollapse.R
\name{Collapses}
\alias{Collapses}
\alias{bsCollapse}
\alias{updateCollapse}
\title{Create Twitter Bootstrap Collapses}
\usage{
bsCollapse(..., id = NULL, multiple = FALSE, open = NULL)

updateCollapse(session, id, open = NULL, close = NULL, type = NULL)
}
\arguments{
\item{id}{The id to assign to the collapse}

\item{multiple}{Logical indicating whether multiple collapse panels can be
open at once}

\item{open}{A vector \code{bsCollapsePanel} ids indicating which panels should be
open}

\item{session}{The session object passed from shinyServer}

\item{close}{A vector of \code{bsCollapsePanel} ids that you want closed.}

\item{\dots}{\code{bsCollapsePanel} elements that will be in the collapse}
}
\value{
If the \code{value} argument was provided to the currently open
\code{\link{bsCollapsePanel}} then that value be returned to the server, otherwise
the \code{\link{bsCollapsePanel}}'s id will be returned.

If \code{multiple == TRUE} then a vector of values or ids will be returned
representing all open \code{\link{bsCollapsePanel}}s.
}
\description{
Functions to create and update Twitter Bootstrap Collapses in shiny
}
\details{


Collapse panels offer an additional way to organize information
within your shiny app. They can be set so that only one panel is open at a
time or can be set so that each panel act independently.

\code{bsCollapse} is a wrapper that links the \code{bsCollapsePanel}s
within it.

Use \code{updateCollapse} in server.R to open and close
\code{bsCollapsePanels} or change their style (danger, info, warning, success).
}
\note{
Run \code{bsDemo()} for a live example of shinyBS Collapses functionality.
}
\section{Options}{
 If \code{multiple = FALSE} (the default), then opening one
\code{\link{bsCollapsePanel}} will close all other \code{\link{bsCollapsePanel}}s within
the same group. If \code{multiple = TRUE}, opening \code{bsCollapsePanel}s
will have no effect on other collapse panels.

\code{open} should be the id of one of the \code{\link{bsCollapsePanel}}s within
the group. If \code{multiple = TRUE}, then it can be a vector of ids.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/}{Twitter Bootstrap 2.3.2}
}
\seealso{
Other collapse: \code{\link{Collapse Panels}},
  \code{\link{bsCollapsePanel}}
}

