% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createAlert.R
\name{createAlert}
\alias{createAlert}
\title{createAlert}
\usage{
createAlert(session, anchorId, alertId = NULL, title = NULL,
  content = NULL, style = NULL, dismiss = TRUE, append = TRUE)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{anchorId}{The unique identifier of the anchor where the alert should be
displayed.}

\item{alertId}{\bold{Optional} A unique identifier for the Alert.}

\item{title}{\bold{Optional} A title for the Alert.}

\item{content}{The main body of the Alert. HTML tags are allowed.}

\item{style}{A bootstrap style to apply. Defaults to \code{info}.}

\item{dismiss}{\bold{logical} Should the Alert be user dismissable? Defaults to \code{TRUE}.}

\item{append}{\code{logical} Should the Alert be appended below existing Alerts? Default to \code{TRUE}.}
}
\description{
\code{createAlert} is used within the Server logic of your Shiny app to display
an alert to the user.
}
\details{
See \link{Alerts} for more information about how to use \code{createAlert} with the
rest of the Alerts family.
}
\note{
Run \code{example("Alerts", "shinyBS", ask = FALSE)} for an example
of createAlert functionality
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Alerts: \code{\link{Alerts}}; \code{\link{bsAlert}};
  \code{\link{closeAlert}}
}

