% Generated by roxygen2 (4.0.1): do not edit by hand
\name{demos}
\alias{bsDemo}
\alias{demos}
\title{shinyBS Demos}
\usage{
bsDemo(example = "demo", port = NULL,
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"), display.mode = "normal")
}
\arguments{
\item{example}{The demo you want to run, currently either 'demo' or 'menubar'.}

\item{port}{The TCP port that the application should listen on. Defaults to
choosing a random port.}

\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only. This value of this parameter can also be a
function to call with the application's URL.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{display.mode}{The mode in which to display the application. If set to
  the value \code{"showcase"}, shows application code and metadata from a
  \code{DESCRIPTION} file in the application directory alongside the
  application. If set to \code{"normal"}, displays the application normally.
  Defaults to \code{"auto"}, which displays the application in the mode
  given in its \code{DESCRIPTION} file, if any.}
}
\description{
Open demonstrations of shinyBS functionality
}
\details{
A wrapper for shiny's \code{\link{runApp}} that makes running shinyBS
 demos easier. Similiar to shiny's \code{\link{runExample}}.

 See \code{\link{runApp}} and \code{\link{runExample}} from shiny for more
 details.
}
\author{
Eric Bailey
}
\seealso{
Other shinyBS_utilities: \code{\link{Remote Shiny Apps}},
  \code{\link{close.rApp}}, \code{\link{closeRemoteApp}},
  \code{\link{runRemoteApp}}
}

