\name{bsNavDivider}
\alias{bsNavDivider}
\title{
  Twitter Bootstrap NavBar Divider
}
\description{
  Adds a vertical line to navbar to help set off other navbar elements
}
\usage{
bsNavDivider()
}

\details{
  Works as an \link{actionButton} but can also be used as a regular web link.
}

\references{
\href{http://getbootstrap.com/2.3.2/components.html\#navbar}{Navbars for Twitter Bootstrap 2.3.2}
}
\author{
Eric Bailey
}

\examples{
  \dontrun{Coming Soon}
}