% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bsGlyph}
\alias{bsGlyph}
\title{Glyphicons}
\usage{
bsGlyph(name, white = FALSE)
}
\arguments{
\item{name}{The name of the glyphicon}

\item{white}{Logical indicating whether the white version should be used}
}
\description{
Add Glyphicons to shiny (DEPRECIATED)
}
\details{
Since writing this function, RStudio has added an \code{\link{icon}}
 function to shiny. Use \code{\link{icon}} instead.
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/base-css.html}{Base CSS for
 Twitter Bootstrap 2.3.2}
}
\seealso{
Other depreciated: \code{\link{bsNavBar}},
  \code{\link{bsNavButton}}, \code{\link{bsNavDateInput}},
  \code{\link{bsNavDateRangeInput}},
  \code{\link{bsNavDivider}}, \code{\link{bsNavDropDown}},
  \code{\link{bsNavLink}}, \code{\link{bsNavTextInput}},
  \code{\link{bsNavTextOutput}},
  \code{\link{bsNavToggleLink}},
  \code{\link{bsNavTypeAhead}}, \code{\link{navbar}},
  \code{\link{pageWithNavBar}},
  \code{\link{updateDropDown}}, \code{\link{updateNavBar}},
  \code{\link{updateToggleLink}}
}

