% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Menu Bar}
\alias{Menu Bar}
\alias{bsMenuBar}
\alias{updateMenuBar}
\title{Twitter Bootstrap NavBar}
\usage{
bsMenuBar(id, ..., brand, rightItems, fixed = FALSE, inverse = FALSE)

updateMenuBar(session, id, brand = NULL, fixed = NULL, inverse = NULL)
}
\arguments{
\item{id}{The id of the menubar object}

\item{\dots}{Various inputs to add to the menu bar. These will be left-aligned}

\item{brand}{A word or brief phrase that will appear on the left of the menu
bar}

\item{rightItems}{a list of additional inputs that will be right-aligned in
the menu bar}

\item{fixed}{Logical indicating whether the menu bar should be fixed to the
top of the screen}

\item{inverse}{Logical indicating whether the menu bar should use the
'inverse' color scheme}

\item{session}{The \code{session} object passed to function given to
 \code{shinyServer}}
}
\description{
Function for creating a Twitter Bootstrap-style navbar for use
 with shinyBS's menu objects
}
\details{
This function is almost identical to the depreciated \code{bsNavBar}
 function but has been renamed to be consistent with the new naming
 convention. \cr\cr Setting \code{fixed = TRUE} will result in a menubar that
 remains at the top of the screen, even if the user scrolls the page. \cr\cr
 Setting \code{inverse = TRUE} will cause the menubar to be black with white
 text.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/components.html}{Components of
 Twitter Bootstrap 2.3.2}

\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}
}

