% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bsMenuLink}
\alias{Menu Bar Inputs}
\alias{bsMenuBarDivider}
\alias{bsMenuDateRangeInput}
\alias{bsMenuLink}
\alias{bsMenuToggleLink}
\alias{bsMenuWrap}
\alias{updateMenuToggle}
\title{Custom inputs for shinyBS menu bars}
\usage{
bsMenuLink(inputId, label, href = "#", icon = NULL)

bsMenuToggleLink(inputId, label, icon = NULL, value = FALSE)

updateMenuToggle(session, inputId, label = NULL, value = NULL)

bsMenuBarDivider()

bsMenuDateRangeInput(inputId, label, start = NULL, end = NULL, min = NULL,
  max = NULL, format = "yyyy-mm-dd", startview = "month", weekstart = 0,
  language = "en", presets)

bsMenuWrap(input, width = 10, form = "auto", icon, placement = "before")
}
\arguments{
\item{input}{An input object}

\item{width}{The width of a text-based input, in characters}

\item{form}{Logical indicating whether the input should be wrapped in a form
tag (See Details)}

\item{icon}{A \href{http://fortawesome.github.io/Font-Awesome/}{Font Awesome}
icon to attach to the input}

\item{placement}{Where to place the icon, either 'before' or 'after'}

\item{inputId}{An Id for the dateRangeInput object}

\item{label}{Label for dateRangeInput (appears as tooltip)}

\item{start}{See \link{dateRangeInput}}

\item{end}{See \link{dateRangeInput}}

\item{min}{See \link{dateRangeInput}}

\item{max}{See \link{dateRangeInput}}

\item{format}{See \link{dateRangeInput}}

\item{startview}{See \link{dateRangeInput}}

\item{weekstart}{See \link{dateRangeInput}}

\item{language}{See \link{dateRangeInput}}

\item{href}{If provided, bsMenuLink will act as a regular html link to the
provided address}

\item{value}{The value to assign to the toggle link, TRUE or FALSE}

\item{session}{The \code{session} object passed to function given to
 \code{shinyServer}}
}
\description{
Function to create specially formatted inputs for shinyBS menu
 bars
}
\details{
Use \code{bsMenuWrapper} to format standard shiny inputs to fit
 better within a menu bar. Only works for text-based inputs (e.g.
 \code{textInput}, \code{dateInput}, \code{numericInput}, etc.) and buttons.
 For \code{dateRangeInput} use the custom \code{bsMenuDateRangeInput}.
 Because of the presence of two text inputs it was easier to write a custom
 function.\cr\cr These functions remove the normal label from the input and
 make it a placeholder within the input. In addition, if an icon is included,
 a tooltip with the label is attached to the icon on hover.\cr\cr Some inputs
 look better when wrapped in a \code{form} tag. if \code{form} isn't
 specified, \code{bsMenuWrap} attempts to determine if the input should be
 wrapped in a \code{form}. If an input is not rendering correctly try setting
 \code{form} to \code{TRUE} or \code{FALSE}. Even if this fixes it, please
 file a bug report or send me an email so I fix the problem for others.
}
\note{
Run \code{bsDemo("menubar")} for a live examples of shinyBS menubar
 functionality.
}
\examples{
#Run bsDemo("menubar") for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/components.html}{Components of
 Twitter Bootstrap 2.3.2}

\href{http://getbootstrap.com/2.3.2/base-css.html}{Base CSS for
 Twitter Bootstrap 2.3.2}
}

