% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Menus}
\alias{Menus}
\alias{bsMenu}
\title{Create Twitter Bootstrap-style Menus and dropdowns}
\usage{
bsMenu(menuId, ..., label, icon = NULL, type = "toolbar", caret = TRUE,
  target = NULL)
}
\arguments{
\item{menuId}{An Id for the menu}

\item{\dots}{Items to add to the menu}

\item{label}{The label for the menu}

\item{icon}{An optional Font Awesome icon}

\item{type}{The type of toolbar to create}

\item{caret}{Logical indicating whether to include a caret on a toolbar menu}

\item{target}{Id of a target for the popup menu}
}
\description{
Create Twitter Bootstrap-style Menus and dropdowns
}
\details{
Function to create a menu based on Twitter Bootstrap styles.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\section{Options}{
 A menu can take any of three forms determined by the
 \code{type} argument. \code{type = 'toolbar'} is for creating top-level
 menus in a \code{\link{bsMenuBar}} object similar to the depreciated
 \code{\link{bsNavToolbar}}.

 \code{type = 'popup'} will create a context menu accessible by
 right-clicking on the element identified by the \code{target} attribute.

 \code{type = 'submenu'} will create submenus within either \code{toolbar} or
 \code{popup} menus.

 \code{label} is the label to apply to the menu, it is ignored when
 \code{type = 'popup'}.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/components.html}{Components of
 Twitter Bootstrap 2.3.2}

\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}
}

