% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Modals}
\alias{Modals}
\alias{bsModal}
\alias{toggleModal}
\title{Twitter Bootstrap Modal Windows}
\usage{
bsModal(id, title, trigger, ..., href)

toggleModal(session, modalId)
}
\arguments{
\item{id}{Id to assign to the modal window}

\item{modalId}{Id of the modal you want to open/close}

\item{title}{The title of the modal window}

\item{trigger}{The id of the button or anchor that will open the modal window}

\item{\dots}{Elements to add to the body of the modal window}

\item{href}{Location of an html file that will provide content of modal body}

\item{session}{The \code{session} object passed to function given to
 \code{shinyServer}}
}
\description{
Functions for creating, showing, and hiding Twitter Bootstrap
 modal windows in shiny
}
\details{
Modals can contain any combination of html, shiny inputs, or shiny
 outputs.\cr\cr \code{toggleModal} will open the model if it is closed, or
 close it if it is open.\cr\cr
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}
}

