% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bsNavDropDown}
\alias{bsNavBar}
\alias{bsNavButton}
\alias{bsNavDateInput}
\alias{bsNavDateRangeInput}
\alias{bsNavDivider}
\alias{bsNavDropDown}
\alias{bsNavLink}
\alias{bsNavTextInput}
\alias{bsNavTextOutput}
\alias{bsNavToggleLink}
\alias{bsNavTypeAhead}
\alias{navbar}
\alias{pageWithNavBar}
\alias{updateDropDown}
\alias{updateNavBar}
\alias{updateToggleLink}
\title{Depreciated Navbar Functions}
\usage{
bsNavDropDown(inputId, label, choices, selected = "")

updateDropDown(session, inputId, label = NULL, choices = NULL,
  selected = NULL)

bsNavBar(inputId, brand, ..., rightItems, fixed = FALSE, inverse = FALSE)

updateNavBar(session, inputId, brand = NULL, fixed = NULL, inverse = NULL)

pageWithNavBar(title = "", navbar, mainContent)

bsNavTextOutput(outputId)

bsNavLink(inputId, label, href = "#")

bsNavToggleLink(inputId, label, value = FALSE)

updateToggleLink(session, inputId, label = NULL, value = NULL)

bsNavDivider()

bsNavButton(inputId, label)

bsNavTextInput(inputId, label, value = "", width = NULL)

bsNavDateInput(inputId, label, value = NULL, min = NULL, max = NULL,
  format = "yyyy-mm-dd", startview = "month", weekstart = 0,
  language = "en", width = NULL)

bsNavDateRangeInput(inputId, label, start = NULL, end = NULL, min = NULL,
  max = NULL, format = "yyyy-mm-dd", startview = "month", weekstart = 0,
  language = "en", width = NULL)

bsNavTypeAhead(inputId, label, value = "", choices, items = 8,
  minLength = 1, width = NULL)
}
\arguments{
\item{inputId}{The Id of the input}

\item{brand}{A word or short phrase that describes the NavBar}

\item{\dots}{Elements that will appear in the NavBar(left-aligned)}

\item{rightItems}{Elements that will appear in the NavBar(right-aligned)}

\item{fixed}{Logical indicating whether the NavBar should be fixed to the top}

\item{inverse}{Logicial indicating whether the NavBar should have the
'inverse' style applied}

\item{title}{The title of the page}

\item{navbar}{A NavBar object to add to the page}

\item{mainContent}{The main content of the page}

\item{outputId}{The Id of the output}

\item{label}{The label the user will see for the input}

\item{href}{An optional link that will be followed if the user clicks the link
object}

\item{value}{The value the object should return}

\item{width}{The width, in pixels, the object should have}

\item{min}{See \link{dateRangeInput}}

\item{max}{See \link{dateRangeInput}}

\item{format}{See \link{dateRangeInput}}

\item{startview}{See \link{dateRangeInput}}

\item{weekstart}{See \link{dateRangeInput}}

\item{language}{See \link{dateRangeInput}}

\item{start}{See \link{dateRangeInput}}

\item{end}{See \link{dateRangeInput}}

\item{choices}{A list of choices to populate the TypeAhead or Dropdown}

\item{items}{See \link{bsTypeAhead}}

\item{minLength}{See \link{bsTypeAhead}}

\item{selected}{The initial value of the \code{bsNavDropDown}}

\item{session}{The \code{session} object passed to function given to
 \code{shinyServer}}
}
\description{
Depreciated functions related to creating NavBars in shiny. See
 \link{bsMenuBar} for updated functions.
}
\details{
These functions have been depreciated. I have reworked how navbars
work and added new functionality. These new function begin with \code{bsMenu}.
The bsNav function remain to avoid breaking older apps but should not be used
for new apps.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/components.html}{Components of
 Twitter Bootstrap 2.3.2}

\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}

\href{http://getbootstrap.com/2.3.2/base-css.html}{Base CSS for
 Twitter Bootstrap 2.3.2}
}
\seealso{
Other depreciated: \code{\link{bsGlyph}}
}

