% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pageWithMenuBar}
\alias{pageWithMenuBar}
\title{Create a page with a menubar}
\usage{
pageWithMenuBar(title = "", menubar, ...)
}
\arguments{
\item{title}{The title to appear in the browser tab}

\item{menubar}{a menubar object created with \code{\link{bsMenuBar}}}

\item{\dots}{Various input to add to the main content area of the shiny app.}
}
\description{
Create a Shiny UI that contains a shinyBS menubar and an area for the main content
}
\details{
This will create a shiny UI taht can be passed to \code{\link{shinyUI}} inside a ui.R script.
}
\examples{
\dontrun{
 shinyUI(pageWithMenuBar(title = "Example",
         bsMenuBar(
           bsMenuLink("test", "Test"),
         ),
         textOutput("title"),
         plotOutput("plot")
 ))
}
}
\author{
Eric Bailey
}

