% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Popovers}
\alias{Popovers}
\alias{addPopover}
\alias{bsPopover}
\alias{removePopover}
\title{Twitter Bootstrap Popovers}
\usage{
bsPopover(id, title, content, placement = "right", trigger = "click")

addPopover(session, id, title = "", content, placement = "right",
  trigger = "click")

removePopover(session, id)
}
\arguments{
\item{id}{The \code{inputId} of the object to add the popover to}

\item{title}{The title of the popover}

\item{content}{The main content of the popover}

\item{placement}{Where the popover should appear in relation to the object.
Acceptable values are \code{bottom}, \code{top}, \code{left}, and
\code{right}}

\item{trigger}{The action that will make the popover appear. Common values are
\code{hover} or \code{click}}

\item{session}{The \code{session} object passed to function given to
 \code{shinyServer}}
}
\description{
Functions to add and manipulate Twitter Bootstrap Tooltips in
 shiny
}
\details{
\code{bsPopover} is called from \code{ui.R} and will add a
 \code{<script>} element to the page that adds a popover on page load. \cr\cr
 \code{addPopover} is called from \code{server.R} and uses shiny's
 \code{customMessageHandler} to add a popover after page load. \cr\cr When
 the popover is added to the app, a "#" is prepended to the \code{id}
 provided. If you have knowledge of CSS/jQuery you can add additional
 selectors after the id to give finer control over the placement of the
 popover.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}
}

