% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Progress Bars}
\alias{Progress Bars}
\alias{bsProgressBar}
\alias{updateProgressBar}
\title{Create Twitter Bootstrap Progress Bars}
\usage{
bsProgressBar(inputId, value = 0, visible = TRUE, color, striped = FALSE,
  animate = FALSE)

updateProgressBar(session, inputId, value = NULL, visible = NULL,
  color = NULL, striped = NULL, animate = NULL)
}
\arguments{
\item{inputId}{The id to assign to the progress bar}

\item{value}{numeric value between 0 and 100. Indicated percentage filled}

\item{visible}{Logical indicating whether the progress bar should be visible}

\item{color}{The color of the progress bar}

\item{striped}{Logical indicating whether the progress bar shoudl be striped}

\item{animate}{Logical indicating whether the progress bar's stripes should
 move}

\item{session}{The session object passed from shinyServer}
}
\description{
Functions to create and update Twitter Bootstrap Progress Bars in shiny
}
\details{


Progress Bars can be used to give the user feedback on their progress
 through a multistep process or the progress of some long-running background
 process running on the server.

 To use a progress bar effectively you need to know the endpoint that you are
 working towards. That can be difficult unless you are updating based on the
 progress of a loop or *apply function.

\code{updateProgressBar} is called from server.R to change the style
 of the progres bar or change its value.
}
\note{
Run \code{bsDemo()} for a live example of shinyBS Progress Bars functionality.
}
\section{Options}{
 \code{value} should be a numeric value between 0 and 100
 where 0 indicates the bar is empty and 100 indicates the bar is completely
 full.

 \code{visible} makes it easy to hide the progress bar when it isn't in use.

 \code{color} is one of Twitter Bootstrap's predefined color schemes:
 \code{danger}, \code{warning}, \code{success}, or \code{info}. If missing,
 the progress bar will be the default blue color scheme.

 \code{striped = TRUE} will cause the progress bar to have diagonal stripes.

 \code{animate = TRUE} will cause the progress bar to have animated diagonal
 stripes.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/}{Twitter Bootstrap 2.3.2}
}

