% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removePopover.R
\name{removePopover}
\alias{removePopover}
\title{removePopover}
\usage{
removePopover(session, id)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{id}{The id of the element to remove the popover from.}
}
\description{
\code{removePopover} is used within the Server logic of an app to remove an
existing popover from a Shiny input or output.
}
\details{
See \link{Tooltips_and_Popovers} for more information about how to use \code{removePopover} with the
rest of the Tooltips_and_Popovers family.
}
\note{
Run \code{bsExample("Tooltips_and_Popovers")} for an example
of \code{removePopover} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Tooltips_and_Popovers: \code{\link{Tooltips_and_Popovers}};
  \code{\link{addPopover}}; \code{\link{addTooltip}};
  \code{\link{bsPopover}}; \code{\link{bsTooltip}};
  \code{\link{popify}}; \code{\link{removeTooltip}};
  \code{\link{tipify}}
}

