\name{runBSExample}
\alias{runBSExample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
shinyBS Demo
}
\description{
Run a shiny app demoing the functionality of the shinyBS package
}
\usage{
runBSExample()
}
\author{
Eric Bailey}
\references{
 Twitter Bootstrap Homepage
}

\examples{
  runBSExample()
}