% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runRemoteApp}
\alias{close.rApp}
\alias{closeRemoteApp}
\alias{runRemoteApp}
\title{Run shiny apps in seperate processes}
\usage{
runRemoteApp(appDir = getwd(), launch.browser = TRUE, ...)

\method{close}{rApp}(con, ...)

closeRemoteApp(rApp)
}
\arguments{
\item{appDir}{The directory of the application}

\item{launch.browser}{Logical indicating whether the default browser should be
automatically opened.}

\item{\dots}{In \code{runRemoteApp} represents other options to pass to
\link{runApp}. For \code{close} serves no purpose}

\item{rApp}{The object created by runRemoteApp. Used to close app afterwards}

\item{con}{The rApp object created by runRemoteApp}
}
\description{
Open a shiny app in a seperate R processes (Windows Only)
}
\details{
These functions were created to assist in testing with testthat and
 RSelenium packages. This allows the shiny test apps to be opened in a
 seperate process so the main process can be used by testthat.\cr\cr The
 function opens an instance of Rscript.exe and opens a socket to that
 process. Based roughly on how the snow package opens additional instances
 for parallel processing.
}
\examples{
\dontrun{
## Open an app on a seperate process
app <- runRemoteApp()

## Close the remote app
closeRemoteApp(app)
}
}
\author{
Eric Bailey
}

