% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{shinyBS-package}
\alias{shinyBS}
\alias{shinyBS-package}
\title{Twitter Bootstrap Components for Shiny}
\description{
shinyBS makes available many of the the components that are part of Twitter
Bootstap but aren't included in Shiny.
}
\details{
Most components have update functions that allow you to modify the elements
from server.R. All functions are completely documented and
\code{\link{bsDemo}} offers live, interactive examples of all components.

shinyBS allows for the creation of many different Twitter Bootstrap elements.
For example, you can create tooltips and popovers to convey extra information
about elements of your app. Alerts let you share information with your users
about what is happening in the background or problems that may exist.

shinyBS's menu functions allow for the creation of menus based on Twitter
Bootstrap's navbar, or context-sensitive popup menus that appear when the
user right-clicks on a targeted element. These menus can contain command
links that act like action buttons, toggle links that act like checkbox
inputs, or radio links that act like radio buttons.

I've created a website for shinyBS at
\url{http://ebailey78.github.io/shinyBS} that (will have/has) examples and
references for all functions in the package as well as instructions for
installing and getting more help.
}

