% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Tables}
\alias{Tables}
\alias{highlightCells}
\alias{highlightRows}
\title{Table Highlighting}
\usage{
highlightCells(session, id, min = NULL, max = NULL, regex = NULL,
  column = NULL, class = NULL, style = NULL, reset = NULL,
  skip.row.names = TRUE)

highlightRows(session, id, column, min = NULL, max = NULL, regex = NULL,
  class = NULL, style = NULL, reset = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}}

\item{id}{The id of the table you want to highlight}

\item{min}{Minimum value to highlight}

\item{max}{Maximum value to highlight}

\item{regex}{A regular expression to select cells/rows to highlight}

\item{column}{Optional column identifier to limit highlighting}

\item{class}{The class to apply to the highlighted cells}

\item{style}{CSS style to apply to the highlighted cells}

\item{reset}{Logical indicating whether previous highlights should be removed
when this highlight is applied}

\item{skip.row.names}{Logical indicating whether the first column should be
 skipped}
}
\description{
Functions to highlight table cells and/or rows based on their
 contents.
}
\details{
\code{highlightCells()} will highlight individual cells that match
criteria provided by other arguments. \cr\cr \code{highlightRows()} will
highlight rows based on any cell in that row matching the provided criteria.
\cr\cr Setting \code{column} will limit any matching to only the column
specified. \cr\cr \code{class} can be set to any of the following:
\code{danger}, \code{info}, \code{warning}, \code{success}, \code{good},
\code{bad}, or \code{neutral}.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/}{Twitter Bootstrap 2.3.2}
}

