% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipify.R
\name{tipify}
\alias{tipify}
\title{tipify}
\usage{
tipify(el, title, placement = "bottom", trigger = "hover", options = NULL)
}
\arguments{
\item{el}{A shiny UI element.}

\item{title}{The content of the tooltip.}

\item{placement}{Where the tooltip should appear relative to its target 
(\code{top}, \code{bottom}, \code{left}, or \code{right}). Defaults to \code{"bottom"}.}

\item{trigger}{What action should cause the tooltip to appear? (\code{hover},
\code{focus}, \code{click}, or \code{manual}). Defaults to \code{"hover"}.}

\item{options}{A named list of additional options to be set on the tooltip.}
}
\description{
\code{tipify} can be wrapped around any shiny UI element to add a tooltip to the
wrapped element. This should be a safer way to add tooltips to elements created with
\code{\link{renderUI}}.
}
\details{
See \link{Tooltips_and_Popovers} for more information about how to use \code{tipify} with the
rest of the Tooltips_and_Popovers family.
}
\note{
Run \code{bsExample("Tooltips_and_Popovers")} for an example
of \code{tipify} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Tooltips_and_Popovers: \code{\link{Tooltips_and_Popovers}},
  \code{\link{addPopover}}, \code{\link{addTooltip}},
  \code{\link{bsPopover}}, \code{\link{bsTooltip}},
  \code{\link{popify}}, \code{\link{removePopover}},
  \code{\link{removeTooltip}}
}

