% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Tooltips}
\alias{Tooltips}
\alias{addTooltip}
\alias{bsTooltip}
\alias{removeTooltip}
\title{Twitter Bootstrap Tooltips}
\usage{
bsTooltip(id, title, placement = "bottom", trigger = "hover")

removeTooltip(session, id)

addTooltip(session, id, title, placement = "bottom", trigger = "hover")
}
\arguments{
\item{id}{The \code{inputId} of the object to add the tooltip to}

\item{title}{The text of the tooltip}

\item{placement}{Where the tooltip should appear in relation to the object.
Acceptable values are \code{bottom}, \code{top}, \code{left}, and
\code{right}}

\item{trigger}{The action that will make the tooltip appear. Common values are
\code{hover} or \code{click}}

\item{session}{The \code{session} object passed to function given to
 \code{shinyServer}}
}
\description{
Functions to add and manipulate Twitter Bootstrap Tooltips in
 shiny
}
\details{
\code{bsTooltip} is called from \code{ui.R} and will add a
 \code{<script>} element to the page that adds a tooltip on page load. \cr\cr
 \code{addTooltip} is called from \code{server.R} and uses shiny's
 \code{customMessageHandler} to add a tooltip after page load. \cr\cr When
 the tooltip is added to the app, a "#" is prepended to the \code{id}
 provided. If you have knowledge of CSS/jQuery you can add additional
 selectors after the id to give finer control over the placement of the
 tooltip.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}
}

