% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TypeAheads}
\alias{TypeAheads}
\alias{bsTypeAhead}
\alias{updateTypeAhead}
\title{Create Twitter Bootstrap TypeAheads}
\usage{
bsTypeAhead(inputId, label, value = "", choices, items = 8, minLength = 1)

updateTypeAhead(session, inputId, label = NULL, value = NULL,
  choices = NULL)
}
\arguments{
\item{inputId}{Id to assign to the typeAhead}

\item{label}{The label that will appear on the typeAhead}

\item{value}{The value to apply to the typeAhead}

\item{choices}{A list or vector of choices that will appear as the user types}

\item{items}{The maximum number of typeAhead choices that will appear at one
time as the user types}

\item{minLength}{The minimum number of characters that a user must type before
typeAhead suggestions begin}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}}
}
\description{
Functions to create and update Twitter Bootstrap TypeAheads in shiny
}
\details{


TypeAheads can be an alternative to select inputs. They appear to be
 a regular text input until the user begins to type. As the user types
 suggestions will appear below the input similar to how Google's search
 suggestions work.

updateTypeAhead is used in server.R to change the style and values of a typeAhead.
}
\note{
Run \code{bsDemo()} for a live example of shinyBS TypeAheads functionality.
}
\section{Options}{

\code{value} should be equal to one of the options presented in choices, but it isn't required.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/}{Twitter Bootstrap 2.3.2}
}

