% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsCollapse.R
\name{updateCollapse}
\alias{updateCollapse}
\title{updateCollapse}
\usage{
updateCollapse(session, id, open = NULL, close = NULL, style = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{id}{The id of the \code{\link{bsCollapse}} object you want to change.}

\item{open}{A vector of \code{value} (or \code{title} if no \code{value} was provided)
of the \code{\link{bsCollapsePanel}} elements you want to open.}

\item{close}{A vector of \code{value} (or \code{title} if no \code{value} was provided)
of the \code{\link{bsCollapsePanel}} elements you want to close.}

\item{style}{A named list of Bootstrap styles (danger, info, warning, success, default).
The names should correspond the the \code{value} of the \code{\link{bsCollapsePanel}} you
want to change.}
}
\description{
Open or close collapse panels or change their styles. See \code{\link{Collapses}}
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}
}
\seealso{
Other Collapses: \code{\link{Collapses}};
  \code{\link{bsCollapsePanel}}; \code{\link{bsCollapse}}
}

