% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTypeahead.R
\name{updateTypeahead}
\alias{updateTypeahead}
\title{updateTypeahead}
\usage{
updateTypeahead(session, inputId, label = NULL, value = NULL, choices = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{Input variable to assign the control's value to}

\item{label}{Display label for the control}

\item{value}{Initial value}

\item{choices}{Array of strings to match against. Can also be JavaScript
function. Use htmlwidgets::JS() to indicate JavaScript.}
}
\description{
\code{updateTypeahead} Update a typeaheadInput buttons.
}
\seealso{
\code{\link{typeaheadInput}}
}
