% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Alerts}
\alias{Alerts}
\alias{bsAlert}
\alias{closeAlert}
\alias{createAlert}
\title{Twitter Bootstrap Alerts}
\usage{
bsAlert(inputId)

createAlert(session, inputId, alertId = NULL, title = NULL,
  message = NULL, type = NULL, dismiss = TRUE, block = NULL,
  append = TRUE)

closeAlert(session, alertId)
}
\arguments{
\item{inputId}{id of the alert anchor}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}}

\item{alertId}{An id for the created alert. Optional, but necessary if you
want to close alert later with \code{closeAlert}}

\item{title}{A title for the alert. This text will be at the top of the alert
and be in larger, bold font}

\item{message}{The main text of the alert}

\item{type}{One of four options: \code{warning}, \code{danger}, \code{info},
or \code{success}. Corresponds to standard Twitter Boostrap styles}

\item{dismiss}{Logical indicating whether the user should be able to close the
alert}

\item{block}{Logical indicating whether 'block' formatting should be applied
to the alert}

\item{append}{Logical indicating whether the alert should be appended to the
 alert anchor or should overwrite existing alerts in the alert anchor}
}
\description{
Functions to add and manipulate Twitter Bootstrap alerts in shiny
}
\details{
By default each new alert will be added below any previously existing
 alerts. If you want only one alert to appear at a time change append to
 FALSE. This will cause new alerts to overwrite any previously existing
 alerts.

 Block style is an option discussed in the Twitter Bootstrap documentation,
 though it doesn't make any difference that the author can see.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/components.html}{Components of
 Twitter Bootstrap 2.3.2}

\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}
}

