% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsCollapse.R
\name{bsCollapse}
\alias{bsCollapse}
\title{bsCollapse}
\usage{
bsCollapse(..., id = NULL, multiple = FALSE, open = NULL)
}
\arguments{
\item{\dots}{\code{\link{bsCollapsePanel}} elements to include in the Collapse.}

\item{id}{\bold{Optional} You can use \code{input$id} in your Server logic to 
determine which panels are open, and \code{\link{updateCollapse}} to open/close
panels.}

\item{multiple}{Can more than one panel be open at a time? Defaults to \code{FALSE}.}

\item{open}{The \code{value}, (or if none was supplied, the \code{title}) of 
the panel(s) you want open on load.}
}
\description{
\code{bsCollapse} is used in your UI to create a collapse panel group. Use
\code{\link{bsCollapsePanel}} to populate this object with panels.
}
\details{
See \link{Collapses} for more information about how to use \code{bsCollapse} with the
rest of the Collapses family.
}
\note{
Run \code{bsExample("Collapses")} for an example
of \code{bsCollapse} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Collapses: 
\code{\link{Collapses}},
\code{\link{bsCollapseButton}()},
\code{\link{bsCollapsePanel}()},
\code{\link{updateCollapse}()}
}
\concept{Collapses}
