% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsCollapsePanel.R
\name{bsCollapsePanel}
\alias{bsCollapsePanel}
\title{bsCollapsePanel}
\usage{
bsCollapsePanel(title, ..., value = title, style = NULL)
}
\arguments{
\item{title}{The title to display at the top of the panel.}

\item{value}{\bold{Optional} The value to return when this panel is open. Defaults to \code{title}.}

\item{style}{\bold{Optional} A Bootstrap style to apply to the panel. (\code{primary}, \code{danger}, \code{warning}, \code{info}, or \code{success})}

\item{\dots}{UI elements to include within the panel.}
}
\description{
\code{bsCollapsePanel} creates individual panels within a \code{\link{bsCollapse}} object.
}
\details{
See \link{Collapses} for more information about how to use \code{bsCollapsePanel} with the
rest of the Collapses family.
}
\note{
Run \code{bsExample("Collapses")} for an example
of \code{bsCollapsePanel} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Collapses: \code{\link{Collapses}},
  \code{\link{bsCollapse}}, \code{\link{updateCollapse}}
}

