% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ButtonGroups}
\alias{ButtonGroups}
\alias{bsButtonGroup}
\alias{updateButtonGroup}
\title{Create Twitter Bootstrap ButtonGroups}
\usage{
bsButtonGroup(inputId, ..., label, toggle = "checkbox", style, size,
  value = NULL, disabled = FALSE, block = FALSE, vertical = FALSE)

updateButtonGroup(session, id, toggle = NULL, style = NULL, size = NULL,
  disabled = NULL, value = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}}
}
\description{
Functions to create and update Twitter Bootstrap ButtonGroups in shiny
}
\details{


ButtonGroups are used to group buttons together. When grouped they
 can act collectively.
}
\note{
Run \code{bsDemo()} for a live example of shinyBS ButtonGroups functionality.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/}{Twitter Bootstrap 2.3.2}
}

