% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Buttons}
\alias{Buttons}
\alias{bsActionButton}
\alias{bsButton}
\alias{bsToggleButton}
\alias{updateButton}
\title{Create Twitter Bootstrap Buttons}
\usage{
bsButton(inputId, label, value, style = NULL, size = NULL, block = FALSE,
  disabled = FALSE)

bsToggleButton(inputId, label, value = FALSE, style = NULL, size = NULL,
  block = FALSE, disabled = FALSE)

bsActionButton(inputId, label, style = NULL, size = NULL, block = FALSE,
  disabled = FALSE)

updateButton(session, id, label = NULL, value = NULL, style = NULL,
  size = NULL, block = NULL, disabled = NULL)
}
\arguments{
\item{inputId}{Input variable to assign the control's value to}

\item{label}{The text that will appear within the button}

\item{value}{See details}

\item{style}{The style to apply to the button}

\item{size}{The desired size of the button}

\item{block}{Logical indicating whether the button should be a block level
control.}

\item{disabled}{Logical indicating whether the button should be clickable}

\item{session}{The session object passed from shinyServer}

\item{id}{The id of the button you want to update with \code{updateButton}}
}
\description{
Functions to create and update Twitter Bootstrap Buttons in shiny
}
\details{


shinyBS adds additional styling options to the
 \code{\link{actionButton}} and introduces two new types of buttons to the
 shiny interface.

\code{bsButton} is a generic button that doesn't have any default
 behavior. They are used as inputs when creating a
 \code{\link{bsButtonGroup}}. They can also be used as triggers for other
 elements like \code{\link{bsModal}} or \code{\link{bsPopover}}.

 For \code{bsButton}, \code{value} is used to assign a return value for
 \code{\link{bsButtonGroup}}. See \code{\link{bsButtonGroup}} for more
 details. For \code{bsToggleButton} and \code{updateButton}, \code{value} is
 used to set whether the button is currently 'clicked'.

 \code{style} is used to customize the appearence of buttons. There are six
 options for style: \code{primary}, \code{info}, \code{success},
 \code{warning}, \code{danger}, \code{inverse}, \code{link}. Any other value
 will result in a standard button style.

 \code{size} is used to control the size of buttons. \code{size} can be
 \code{large}, \code{small}, or \code{mini}. Any other value will result in a
 standard-sized button.

 Setting \code{block = TRUE} will cause the button to take up the full width
 of its parent element.

 When \code{disabled = TRUE} the button will appear "greyed-out" and clicking
 the button will have no effect on the app.

\code{bsToggleButton} is an alternative to
 \code{\link{checkboxInput}}. It has on and an off states that correspond to
 \code{TRUE/FALSE}. When \code{TRUE} the toggle button will look like it is pressed
 in. When \code{FALSE} it will appear as a normal button. Clicking the toggle
 button will alternate between the two states.

\code{bsActionButton} works just like shiny's default
 \code{\link{actionButton}} but has additional styling options.

updateButton can be used within server.R to change the appearence
 and/or state of any of the shinyBS button types. If a parameter is passed to
 \code{updateButton} that doesn't apply to the button selected, it will be
 ignored (e.g. \code{value} for \code{bsActionButton}).
}
\note{
Run \code{bsDemo()} for a live example of shinyBS Buttons functionality.
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/}{Twitter Bootstrap 2.3.2}
}

