% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Buttons}
\alias{Buttons}
\alias{bsActionButton}
\alias{bsButton}
\alias{bsButtonGroup}
\alias{bsToggleButton}
\alias{updateButton}
\alias{updateButtonGroup}
\title{Twitter Bootstrap Buttons}
\usage{
bsButton(inputId, label, value, style = NULL, size = NULL, block = FALSE,
  disabled = FALSE)

bsToggleButton(inputId, label, value = FALSE, style = NULL, size = NULL,
  block = FALSE, disabled = FALSE)

bsActionButton(inputId, label, style = NULL, size = NULL, block = FALSE,
  disabled = FALSE)

updateButton(session, id, label = NULL, value = NULL, style = NULL,
  size = NULL, block = NULL, disabled = NULL)

bsButtonGroup(inputId, ..., label, toggle = "checkbox", style, size,
  value = NULL, disabled = FALSE, block = FALSE, vertical = FALSE)

updateButtonGroup(session, id, toggle = NULL, style = NULL, size = NULL,
  disabled = NULL, value = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}}

\item{inputId}{Id to assign to the button or button group}

\item{id}{The id of the button/button group you want to update}

\item{\dots}{\code{bsButton()} objects to be added to the button group}

\item{label}{For buttons, the text to appear inside the button. For button
groups, an optional label that will appear above the button group}

\item{toggle}{The type of toggle behavior the button group should have (See
Details)}

\item{style}{The bootstrap style the button(s) should take (See Details)}

\item{size}{The bootstrap size the button(s) should take (See Details)}

\item{block}{Logical indicating whether the button or button group should be a
block level element (i.e., should it span the width of its parent element)}

\item{vertical}{Logical indicating whether the button group should have a
vertical orientation}

\item{value}{The value of the button/button group (See Details)}

\item{disabled}{Logical indicating whether the button(s) should be disabled}
}
\description{
Functions for setting and changing the style, size, and state of
 various buttons in shiny apps.
}
\details{
\code{bsActionButton()} creates an action button that behaves just as
a standard shiny action button does. It has the added functionality of being
able to changed its style and size. It can also be disabled/enabled.\cr\cr
\code{toggle} can take a value of either \code{radio} or \code{checkbox}.
\code{radio} will allow only one button in the button group to be selected at
a time. \code{checkbox} will allow any number of buttons to be selected at a
time. \cr\cr \code{style} can be any of the styles described in the Twitter
Bootstrap 2.3.2 documentation. Acceptable values are currently: primary, info,
success, warning, danger, inverse, or link. Additionally, when calling one of
the update functions, style can be set to \code{default} to return to the
default button style.\cr\cr \code{size} can be any of the sizes described in
the Twitter Bootstrap 2.3.2 documentation. Accepatble values are currently:
large, small, or mini. Additionally, when calling one of the update functions,
style can be set to \code{default} to return to the default size.\cr\cr For
toggle buttons, \code{value} can be \code{TRUE} or \code{FALSE} and
corresponds to whether the button is currently 'clicked.' For \code{bsButton},
\code{value} is used to set the value that will be returned by containing
\code{bsButtonGroup} object when the button is clicked. For button groups,
\code{value} is used to set the current value to be returned by the group and
should correspond to values assigned to buttons contained in the button
group.\cr\cr \code{vertical} and \code{block} for button groups are
experimental. They do not work well together and may not work under all
browsers.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/components.html}{Components of
 Twitter Bootstrap 2.3.2}

\href{http://getbootstrap.com/2.3.2/base-css.html}{Base CSS for
 Twitter Bootstrap 2.3.2}
}

