% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsCollapse.R
\name{Collapses}
\alias{Collapses}
\title{Collapse Panels}
\description{
Collapse panels allow you to reduce clutter in your Shiny app by making
panels of information that open and close with a user's click. Any type of
content can go in a collapse panel. Standard Bootstrap styling options are
available.
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\section{Components}{

\describe{
 \item{\code{\link{bsCollapse}}}{A container for holder the individual panels created by \code{\link{bsCollapsePanel}}.}
 \item{\code{\link{bsCollapsePanel}}}{Creates an individual Collapse Panel that resides within a \code{\link{bsCollapse}}.}
 \item{\code{\link{updateCollapse}}}{Used within your server logic to open/close collapse panels or change their style.}
}
}
\references{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}
}
\seealso{
Other Collapses: \code{\link{bsCollapsePanel}};
  \code{\link{bsCollapse}}; \code{\link{updateCollapse}}
}

