% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Demos}
\alias{Demos}
\alias{bsDemo}
\title{shinyBS Demos}
\usage{
bsDemo(example = "demo", port = NULL,
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"), display.mode = "normal")
}
\arguments{
\item{example}{The demo you want to run, currently either 'demo' or 'menubar'.}

\item{port}{See \link{runApp}}

\item{launch.browser}{See \link{runApp}}

\item{host}{See \link{runApp}}

\item{display.mode}{See \link{runApp}}
}
\description{
Open demonstrations of shinyBS functionality
}
\details{
See /link{runApp} and /link{runExample} from shiny for more details.
}
\author{
Eric Bailey
}

