\name{bsNavDivider}
\alias{bsNavDivider}
\title{
  Twitter Bootstrap Navbar Divider
}
\description{
  Adds a vertical line to navbar to help set off other navbar elements
}
\usage{
bsNavDivider()
}

\references{
\href{http://getbootstrap.com/2.3.2/components.html}{Navbars for Twitter Bootstrap 2.3.2}
}
\author{
Eric Bailey
}

\examples{
\dontrun{

  ##In ui.R
  bsNavDivider()
  }
}