% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bsGlyph}
\alias{bsGlyph}
\title{Glyphicons}
\usage{
bsGlyph(name, white = FALSE)
}
\arguments{
\item{name}{The name of the glyphicon}

\item{white}{Logical indicating whether the white version should be used}
}
\description{
Add Glyphicons to shiny (DEPRECIATED)
}
\details{
Since writing this function, RStudio has added an \code{icon()} function to shiny. Use \code{icon()} instead.
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/base-css.html}{Base CSS for
 Twitter Bootstrap 2.3.2}
}

