% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsTooltipAndPopover.R
\name{Popovers}
\alias{Popovers}
\alias{addPopover}
\alias{bsPopover}
\alias{removePopover}
\title{Create Twitter Bootstrap Popovers}
\usage{
bsPopover(id, title, content, placement = "bottom", trigger = "hover",
  options = NULL)

removePopover(session, id)

addPopover(session, id, title, content, placement = "bottom",
  trigger = "hover", options = NULL)
}
\arguments{
\item{id}{The id of the element you want to attach/remove the popover to/from}

\item{title}{The title to appear at the top of the popover.}

\item{content}{The main body of the popover}

\item{placement}{Where the popover shoud appear relative to its target}

\item{trigger}{What action should cause the popover to appear}

\item{options}{A named list of additional popover options (See Details)}

\item{session}{The session object passed from shinyServer}
}
\description{
Functions to create and update Twitter Bootstrap Popovers in shiny
}
\details{


Popovers are used to convey small amounts of information to the user
of your shiny app. They are usually set to appear on click or hover and can
contain text to help better explain an element of your shiny app.

shiny and shinyBS use jQuery to select elements within the page.
shinyBS will prepend the \code{id} argument with a "#" to tell jQuery to look
for an id, but additional css selectors can be included within the \code{id}
argument. For example, if you wanted to attach a popover to the first header
of a table with an id of "tableEx" you could do something like \code{id =
"tableEx th:nth-child(1)"}.

\code{options} gives access to less commonly used options in the bootstrap
popover API. It should be a named list in which the names match options for
popovers (See the Bootstrap documentation). Values provided in \code{options}
will overwrite any values provided by \code{title}, \code{content},
\code{placement}, or \code{trigger} arguments.

\code{bsPopover} is used from within ui.R to add popovers to elements

\code{removePopover} will remove any popovers attached to the
selected element.

\code{addPopover} is called from within server.R to add popovers to
elements
}
\note{
Run \code{bsDemo()} for a live example of shinyBS Popovers functionality.
}
\section{Options}{
 \code{placement} can be \code{right}, \code{left},
\code{top}, or \code{bottom}

\code{trigger} can be \code{hover}, \code{click}, \code{focus}, or
\code{manual}
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/}{Twitter Bootstrap 2.3.2}
}

