% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Popovers}
\alias{Popovers}
\alias{addPopover}
\alias{bsPopover}
\alias{removePopover}
\title{Create Twitter Bootstrap Popovers}
\usage{
bsPopover(id, title, content, placement = "right", trigger = "click")

addPopover(session, id, title = "", content, placement = "right",
  trigger = "click")

removePopover(session, id)
}
\arguments{
\item{id}{The id of the element you want to attach/remove the popover to/from}

\item{title}{The title to appear at the top of the popover.}

\item{content}{The main body of the popover}

\item{placement}{Where the popover shoud appear relative to its target}

\item{trigger}{What action should cause the popover to appear}

\item{session}{The session object passed from shinyServer}
}
\description{
Functions to create and update Twitter Bootstrap Popovers in shiny
}
\details{


Popovers are used to convey small amounts of information to the user
 of your shiny app. They are usually set to appear on click or hover and can
 contain text to help better explain an element of your shiny app.

shiny and shinyBS use jQuery to select elements within the page.
shinyBS will prepend the \code{id} argument with a "#" to tell jQuery to look
for an id, but additional css selectors can be included within the \code{id}
argument. For example, if you wanted to attach a popover to the first header
of a table with an id of "tableEx" you could do something like \code{id =
"tableEx th:nth-child(1)"}.

\code{bsPopover} is used from within ui.R to add
popovers to elements

\code{addPopover} is called from within server.R to add popovers to elements

\code{removePopover} will remove any popovers attached to the selected element.
}
\note{
Run \code{bsDemo()} for a live example of shinyBS Popovers functionality.
}
\section{Options}{
 \code{placement} can be \code{right}, \code{left},
 \code{top}, or \code{bottom}

 \code{trigger} can be \code{hover}, \code{click}, \code{focus}, or
 \code{manual}
}
\examples{
\dontrun{Run bsDemo() for examples}
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/}{Twitter Bootstrap 2.3.2}
}

