% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeTooltip.R
\name{removeTooltip}
\alias{removeTooltip}
\title{removeTooltip}
\usage{
removeTooltip(session, id)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{id}{The id of the element to remove the tooltip from.}
}
\description{
\code{removeTooltip} is used within the Server logic of an app to remove an
existing tooltip from a Shiny input or output.
}
\details{
See \link{Tooltips_and_Popovers} for more information about how to use \code{removeTooltip} with the
rest of the Tooltips_and_Popovers family.
}
\note{
Run \code{example("Tooltips_and_Popovers", "shinyBS", ask = FALSE)} for an example
of removeTooltip functionality
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Tooltips_and_Popovers: \code{\link{Tooltips_and_Popovers}};
  \code{\link{addPopover}}; \code{\link{addTooltip}};
  \code{\link{bsPopover}}; \code{\link{bsTooltip}};
  \code{\link{removePopover}}
}

