% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toggleModal.R
\name{toggleModal}
\alias{toggleModal}
\title{toggleModal}
\usage{
toggleModal(session, modalId, toggle = "toggle")
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{modalId}{The id of the modal window you want to open/close}

\item{toggle}{Should the modal window \code{open}, \code{close}, or \code{toggle}?}
}
\description{
\code{toggleModal} is used within your Server logic to open or close a modal
window.
}
\details{
See \link{Modals} for more information about how to use \code{toggleModal} with the
rest of the Modals family.
}
\note{
Run \code{example("Modals", "shinyBS", ask = FALSE)} for an example
of toggleModal functionality
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Modals: \code{\link{Modals}}; \code{\link{bsModal}}
}

