% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TypeAhead}
\alias{TypeAhead}
\alias{bsTypeAhead}
\alias{updateTypeAhead}
\title{TWitter Bootstrap TypeAheads}
\usage{
bsTypeAhead(inputId, label, value = "", choices, items = 8, minLength = 1)

updateTypeAhead(session, inputId, label = NULL, value = NULL,
  choices = NULL)
}
\arguments{
\item{inputId}{Id to assign to the typeAhead}

\item{label}{The label that will appear on the typeAhead}

\item{value}{The value to apply to the typeAhead}

\item{choices}{A list of choices taht will appear as the user types}

\item{items}{The maximum number of typeAhead choices that will appear as one
time as the user types}

\item{minLength}{The minimum number of characters that a use must type before
typeAhead suggestions begin}

\item{session}{The \code{session} object passed to function given to
 \code{shinyServer}

 #'}
}
\description{
Functions to add and update Twitter Bootstrap TypeAheads in shiny
}
\note{
Run \code{bsDemo()} for a live examples of shinyBS functionality.
}
\examples{
#Run bsDemo() for examples
}
\author{
Eric Bailey
}
\references{
\href{http://getbootstrap.com/2.3.2/javascript.html}{Javascript
 for Twitter Bootstrap 2.3.2}
}

