% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCollapse.R
\name{updateCollapse}
\alias{updateCollapse}
\title{updateCollapse}
\usage{
updateCollapse(session, id, open = NULL, close = NULL, style = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{id}{The id of the Collapse object you want to change.}

\item{open}{A vector of \code{value} (or \code{title} if no \code{value} was 
provided) values identifying the panels you want to open.}

\item{close}{A vector of \code{value} (or \code{title} if no \code{value} was 
provided) values identifying the panels you want to close.}

\item{style}{A named list of Bootstrap styles (\code{primary}, \code{danger}, \code{info},
\code{warning}, \code{success}, or \code{default}). The names should correspond
to the \code{value} (or \code{title} if no \code{value} was provided) of the 
\code{\link{bsCollapsePanel}} you want to change.}
}
\description{
\code{updateCollapse} is used within the Server logic of your Shiny app to
modify a Collapse after load.
}
\details{
See \link{Collapses} for more information about how to use \code{updateCollapse} with the
rest of the Collapses family.
}
\note{
Run \code{bsExample("Collapses")} for an example
of \code{updateCollapse} functionality.
}
\seealso{
\href{http://getbootstrap.com}{Twitter Bootstrap 3}

Other Collapses: 
\code{\link{Collapses}},
\code{\link{bsCollapsePanel}()},
\code{\link{bsCollapse}()}
}
\concept{Collapses}
